refactor: refine handler pkg
